+++ src/briefy/common/db/mixins/__init__.py
@@ -1,8 +1,10 @@
 """Database mixins."""
 from briefy.common.db.mixins.address import Address
+from briefy.common.db.mixins.asset import Image  # noQA
+from briefy.common.db.mixins.asset import ThreeSixtyImage  # noQA
+from briefy.common.db.mixins.asset import Video  # noQA
 from briefy.common.db.mixins.external import KnackMixin  # noQA
 from briefy.common.db.mixins.identifiable import GUID
-from briefy.common.db.mixins.image import Image  # noQA
 from briefy.common.db.mixins.metadata import BaseMetadata  # noQA
 from briefy.common.db.mixins.person import NameMixin  # noQA
 from briefy.common.db.mixins.person import PersonalInfoMixin  # noQA